backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=12.1.1",
    "pystac-client>=0.9.0",
    "rasterio>=1.5.0",
    "requests>=2.32.5",
    "titiler>=0.24.2",
]
